FEAT: Edited myPageInfo edit process
- Added myInfo edit process
