+++ src/Routes/MyPage/index.js
@@ -1,0 +1,3 @@
+import MyPageContainer from "./MyPageContainer";
+
+export default MyPageContainer;
