+++ src/Routes/PostDetail.js
@@ -1,0 +1,1 @@
+export default () => "Post Detail";
